+++ CHANGELOG.md
@@ -7,9 +7,10 @@
       the behavior of
       [dateutil.tz.gettz()](https://dateutil.readthedocs.io/en/stable/tz.html).
     * Rename project and packages
+        * Rename project from `AceTimePython` to `acetimepy` to be more
+          compatible with Python conventions.
         * Rename `zonedb_types.py` to `typing.py`.
         * Rename `acetime.acetz` (acetz.py) to `acetime.timezone` (timezone.py).
-        * Rename project from `AceTimePython` to `acetimepy`.
 * 0.6.1 (2023-04-01, TZDB 2023c)
     * Upgrade TZDB from 2023b to 2023c.
         * https://mm.icann.org/pipermail/tz-announce/2023-March/000079.html