+++ CHANGELOG.md
@@ -15,6 +15,7 @@
       an exception if the `zone_name` is not found. This is more consistent with
       the behavior of
       [dateutil.tz.gettz()](https://dateutil.readthedocs.io/en/stable/tz.html).
+    * Move `utils/zinfo.py` from `AceTimeTools` project.
 * 0.6.1 (2023-04-01, TZDB 2023c)
     * Upgrade TZDB from 2023b to 2023c.
         * https://mm.icann.org/pipermail/tz-announce/2023-March/000079.html